setup template areas
